Added env support also tags are cached in their own folders

+++ src/Artdevue/Fcache/Fcache.php
@@ -1,4 +1,5 @@
 <?php namespace Artdevue\Fcache;
+
 /**
  * Created by PhpStorm.
  * User: AtrDevue
@@ -9,7 +10,8 @@ use Illuminate\Cache\StoreInterface,
     Illuminate\Filesystem\Filesystem,
     Closure;
 
-class Fcache implements StoreInterface {
+class Fcache implements StoreInterface
+{
 
     /**
      * The Illuminate Filesystem instance.
@@ -49,14 +51,14 @@ class Fcache implements StoreInterface {
     /**
      * Create a new Dummy cache store.
      *
-     * @param  string  $prefix
+     * @param  string $prefix
      * @return void
      */
     public function __construct($prefix = '')
     {
         $this->files = new Filesystem;
         $this->prefix = $prefix;
-        $this->directory = \Config::get('cache.path');
+        $this->directory = \Config::get('cache.path') . '/' . app()->environment();
         $this->tags = array();
         $this->directoryTags = $this->directory . (!empty($prefix) ? '/' . $prefix : '') . '/tags';
     }
@@ -64,32 +66,30 @@ class Fcache implements StoreInterface {
     /**
      * Retrieve an item from the cache by key.
      *
-     * @param  string  $key
+     * @param  string $key
      * @return mixed
      */
     public function get($key)
     {
         $path = $this->path($key);
 
-        if ( ! $this->files->exists($path))
-        {
+
+        if (!$this->files->exists($path)) {
+
+
             return null;
         }
 
-        try
-        {
+        try {
             $expire = substr($contents = $this->files->get($path), 0, 10);
-        }
-        catch (\Exception $e)
-        {
+        } catch (\Exception $e) {
             return null;
         }
 
         // If the current time is greater than expiration timestamps we will delete
         // the file and return null. This helps clean up the old files and keeps
         // this directory much cleaner for us as old files aren't hanging out.
-        if (time() >= $expire)
-        {
+        if (time() >= $expire) {
             return $this->forget($key);
         }
 
@@ -99,19 +99,23 @@ class Fcache implements StoreInterface {
     /**
      * Store an item in the cache for a given number of minutes.
      *
-     * @param  string  $key
-     * @param  mixed   $value
-     * @param  int     $minutes
+     * @param  string $key
+     * @param  mixed $value
+     * @param  int $minutes
      * @return void
      */
     public function put($key, $value, $minutes)
     {
-        $value = $this->expiration($minutes).serialize($value);
+        $value = $this->expiration($minutes) . serialize($value);
 
         $this->createCacheDirectory($path = $this->path($key));
 
-        if (!empty($this->tags))
+        if (!empty($this->tags)) {
             $this->_setTags($path);
+        }
+
+        $path = $this->normalize_path($path, false);
+
 
         $this->files->put($path, $value);
     }
@@ -119,14 +123,14 @@ class Fcache implements StoreInterface {
     /**
      * Tags for cache.
      *
-     * @param  string  $string
+     * @param  string $string
      * @return object
      */
     public function tags($tags)
     {
-        if(is_string($tags))
+        if (is_string($tags))
             $string_array = explode(',', $tags);
-        else if(is_array($tags))
+        else if (is_array($tags))
             $string_array = $tags;
         array_walk($string_array, 'trim');
 
@@ -138,39 +142,38 @@ class Fcache implements StoreInterface {
     /**
      * Save Tags for cache.
      *
-     * @param  string  $path
+     * @param  string $path
      * @return null
      */
     private function _setTags($path)
     {
 
-        foreach ($this->tags as $tg)
-        {
-            $file = $this->directoryTags . '/'. $tg;
-            if (!$this->files->exists($file))
-            {
+        foreach ($this->tags as $tg) {
+            $file = $this->directoryTags . '/' . $tg;
+            $file = $this->normalize_path($file, false);
+
+            if (!$this->files->exists($file)) {
                 $this->createCacheDirectory($file);
                 $this->files->put($file, $path);
-            }
-            else
-            {
+
+
+            } else {
                 $farr = file($file);
-                if (!in_array($path, $farr))
-                {
-                    file_put_contents($file,"\n$path", FILE_APPEND);
+                if (!in_array($path, $farr)) {
+                    file_put_contents($file, "\n$path", FILE_APPEND);
                 }
             }
 
         }
-        $this->tags = array();
+        // $this->tags = array();
     }
 
     /**
      * Get an item from the cache, or store the default value.
      *
-     * @param  string  $key
-     * @param  \DateTime|int  $minutes
-     * @param  Closure  $callback
+     * @param  string $key
+     * @param  \DateTime|int $minutes
+     * @param  Closure $callback
      * @return mixed
      */
     public function remember($key, $minutes, Closure $callback)
@@ -178,8 +181,7 @@ class Fcache implements StoreInterface {
         // If the item exists in the cache we will just return this immediately
         // otherwise we will execute the given Closure and cache the result
         // of that execution for the given number of minutes in storage.
-        if ( ! is_null($value = $this->get($key)))
-        {
+        if (!is_null($value = $this->get($key))) {
             return $value;
         }
 
@@ -191,8 +193,8 @@ class Fcache implements StoreInterface {
     /**
      * Get an item from the cache, or store the default value forever.
      *
-     * @param  string   $key
-     * @param  Closure  $callback
+     * @param  string $key
+     * @param  Closure $callback
      * @return mixed
      */
     public function rememberForever($key, Closure $callback)
@@ -200,8 +202,7 @@ class Fcache implements StoreInterface {
         // If the item exists in the cache we will just return this immediately
         // otherwise we will execute the given Closure and cache the result
         // of that execution for the given number of minutes. It's easy.
-        if ( ! is_null($value = $this->get($key)))
-        {
+        if (!is_null($value = $this->get($key))) {
             return $value;
         }
 
@@ -213,17 +214,14 @@ class Fcache implements StoreInterface {
     /**
      * Create the file cache directory if necessary.
      *
-     * @param  string  $path
+     * @param  string $path
      * @return void
      */
     protected function createCacheDirectory($path)
     {
-        try
-        {
+        try {
             $this->files->makeDirectory(dirname($path), 0777, true, true);
-        }
-        catch (\Exception $e)
-        {
+        } catch (\Exception $e) {
             //
         }
     }
@@ -231,8 +229,8 @@ class Fcache implements StoreInterface {
     /**
      * Increment the value of an item in the cache.
      *
-     * @param  string  $key
-     * @param  mixed   $value
+     * @param  string $key
+     * @param  mixed $value
      * @return void
      *
      * @throws \LogicException
@@ -245,8 +243,8 @@ class Fcache implements StoreInterface {
     /**
      * Increment the value of an item in the cache.
      *
-     * @param  string  $key
-     * @param  mixed   $value
+     * @param  string $key
+     * @param  mixed $value
      * @return void
      *
      * @throws \LogicException
@@ -259,8 +257,8 @@ class Fcache implements StoreInterface {
     /**
      * Store an item in the cache indefinitely.
      *
-     * @param  string  $key
-     * @param  mixed   $value
+     * @param  string $key
+     * @param  mixed $value
      * @return void
      */
     public function forever($key, $value)
@@ -271,7 +269,7 @@ class Fcache implements StoreInterface {
     /**
      * Remove an item from the cache by tags.
      *
-     * @param  string  $string
+     * @param  string $string
      * @return void
      */
     public function forgetTags($string)
@@ -279,16 +277,19 @@ class Fcache implements StoreInterface {
         $string_array = explode(',', $string);
         array_walk($string_array, 'trim');
 
-        foreach ($string_array as $sa)
-        {
-            $file = $this->directoryTags . '/'. $sa;
-            if ($this->files->exists($file))
-            {
+        foreach ($string_array as $sa) {
+            $file = $this->directoryTags . '/' . $sa;
+            if ($this->files->exists($file)) {
                 $farr = file($file);
-                foreach ($farr as $f)
-                {
-                    if ($this->files->exists($f))
+                foreach ($farr as $f) {
+                    $f = $this->normalize_path($f, false);
+                    if ($this->files->exists($f)) {
+
                         unlink($f);
+                    } else {
+
+                    }
+
                 }
                 unlink($file);
             }
@@ -298,22 +299,18 @@ class Fcache implements StoreInterface {
     /**
      * Remove an item from the cache.
      *
-     * @param  string  $key
+     * @param  string $key
      * @return void
      */
     public function forget($key)
     {
         $file = $this->path($key);
 
-        if ($this->files->exists($file))
-        {
+        if ($this->files->exists($file)) {
             $this->files->delete($file);
-        }
-        else
-        {
+        } else {
             $folder = substr($file, 0, -6);
-            if ($this->files->exists($folder))
-            {
+            if ($this->files->exists($folder)) {
                 $this->files->deleteDirectory($folder);
             }
         }
@@ -322,14 +319,22 @@ class Fcache implements StoreInterface {
     /**
      * Remove all items from the cache.
      *
-     * @param  string  $tag
+     * @param  string $tag
      * @return void
      */
     public function flush()
     {
-        foreach ($this->files->directories($this->directory) as $directory)
-        {
-            $this->files->deleteDirectory($directory);
+        if (empty($this->tags)) {
+            foreach ($this->files->directories($this->directory) as $directory) {
+                $this->files->deleteDirectory($directory);
+            }
+        }
+
+        foreach ($this->tags as $tag) {
+            $items = $this->forgetTags($tag);
+            $del = $this->directory . '/' . $tag;
+            $del = $this->normalize_path($del);
+            $this->files->deleteDirectory($del);
         }
     }
 
@@ -366,30 +371,26 @@ class Fcache implements StoreInterface {
     /**
      * Get the full path for the given cache key.
      *
-     * @param  string  $key
+     * @param  string $key
      * @return string
      */
     protected function path($key)
     {
-        /*$parts = '';
-        $array = explode('/',$key);
-
-        $ph = array_diff($array, array(end($array)));
-
-        if (count($ph > 0))
-            $parts = implode('/',$ph) . '/';
-
-        return $this->directory.'/'. $parts . end($array) . '.cache';*/
 
         $prefix = !empty($this->prefix) ? $this->prefix . '/' : '';
+        $subdir = 'global';
+        if (!empty($this->tags)) {
+            $subdir = reset($this->tags);
+        }
 
-        return $this->directory.'/'. $prefix . trim($key, "/") . '.cache';
+
+        return $this->directory . '/' . $subdir . '/' . $prefix . trim($key, "/") . '.cache';
     }
 
     /**
      * Get the expiration time based on the given minutes.
      *
-     * @param  int  $minutes
+     * @param  int $minutes
      * @return int
      */
     protected function expiration($minutes)
@@ -398,4 +399,41 @@ class Fcache implements StoreInterface {
 
         return time() + ($minutes * 60);
     }
+
+
+    function normalize_path($path, $slash_it = true)
+    {
+        $path_original = $path;
+        $s = DIRECTORY_SEPARATOR;
+        $path = preg_replace('/[\/\\\]/', $s, $path);
+        $path = str_replace($s . $s, $s, $path);
+        if (strval($path) == '') {
+            $path = $path_original;
+        }
+        if ($slash_it == false) {
+            $path = rtrim($path, DIRECTORY_SEPARATOR);
+        } else {
+            $path .= DIRECTORY_SEPARATOR;
+            $path = rtrim($path, DIRECTORY_SEPARATOR . DIRECTORY_SEPARATOR);
+        }
+        if (strval(trim($path)) == '' or strval(trim($path)) == '/') {
+            $path = $path_original;
+        }
+        if ($slash_it == false) {
+        } else {
+            $path = $path . DIRECTORY_SEPARATOR;
+            $path = $this->reduce_double_slashes($path);
+        }
+
+
+        return $path;
+    }
+
+
+    function reduce_double_slashes($str)
+    {
+        return preg_replace("#([^:])//+#", "\\1/", $str);
+    }
+
+
 }
